requirements.txt
# Core Data Processing
pandas==2.1.4
numpy==1.26.3

# Database Connectivity
pyodbc==5.0.1
SQLAlchemy==2.0.25

# Environment Management
python-dotenv==1.0.0

# Data Quality & Validation
great-expectations==0.18.8

# Workflow Orchestration
apache-airflow==2.8.1

# Testing
pytest==7.4.4
pytest-cov==4.1.0

# Logging & Monitoring
colorlog==6.8.2

# Data Visualization (optional)
matplotlib==3.8.2
seaborn==0.13.1

# Jupyter (for exploration)
jupyter==1.0.0
ipykernel==6.29.0

# Code Quality
black==24.1.1
flake8==7.0.0
pylint==3.0.3
